requirements.txt
flask==3.1.0
pdfplumber==0.11.4
openai==1.41.0
fpdf2==2.8.3
httpx==0.28.1
python-dotenv==1.0.1
